Finished project

+++ AirQuality ObjC/AirQuality-C/AirQuality-C/Models/DVMWeather.h
@@ -14,9 +14,9 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (nonatomic, readonly) NSInteger temperature;
 @property (nonatomic, readonly) NSInteger humidity;
-@property (nonatomic, readonly) float windSpeed;
+@property (nonatomic, readonly) NSInteger windSpeed;
 
--(instancetype)initWithWeatherInfo:(NSInteger)temperature humidity:(NSInteger)humidity windSpeed:(float)windSpeed;
+-(instancetype)initWithWeatherInfo:(NSInteger)temperature humidity:(NSInteger)humidity windSpeed:(NSInteger)windSpeed;
 
 @end
